Refactor calendar components and add ListView component
